+++ CHANGELOG.md
@@ -10,3 +10,5 @@ Change Log
 - fixed multi-page queries when `--num N` was greater that the initial result size
 - added `--np` and `--noprompt` options to just search and exit
 - added `-l` `--language` command line options and `language` option in config file
+- added `-j` `--first` command line option to open the first result and exit
+- added `--lucky` command line option to open a random result and exit